Split touch pointer event into two
For touch type pointers gecko doesn't support all the properties in pointer-events (and
it's not even very clear to me that they all make sense; how can one talk about the twist of
a finger, for example?). So split out a version of the tests that only test the properties
which are supported, so that we can regression-test the basics without failing due to missing
support at the engine layer for more complex properties.

Differential Revision: https://phabricator.services.mozilla.com/D139121

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1543337
gecko-commit: f0192382bd67ab400304eab0bf544a93430cfd79
gecko-reviewers: webdriver-reviewers, whimboo, jdescottes
