+++ webdriver/tests/perform_actions/pointer_mouse.py
@@ -118,85 +118,6 @@ def test_click_navigation(session, url, inline):
     Poll(session, message=error_message).until(lambda s: s.url == destination)
 
 
-def test_pen_pointer_properties(session, test_actions_pointer_page, pen_chain):
-    pointerArea = session.find.css("#pointerArea", all=False)
-    center = get_inview_center(pointerArea.rect, get_viewport_rect(session))
-    pen_chain.pointer_move(0, 0, origin=pointerArea) \
-        .pointer_down(pressure=0.36, tilt_x=-72, tilt_y=9, twist=86) \
-        .pointer_move(10, 10, origin=pointerArea) \
-        .pointer_up() \
-        .pointer_move(80, 50, origin=pointerArea) \
-        .perform()
-    events = get_events(session)
-    assert len(events) == 10
-    event_types = [e["type"] for e in events]
-    assert ["pointerover", "pointerenter", "pointermove", "pointerdown",
-            "pointerover", "pointerenter", "pointermove", "pointerup",
-            "pointerout", "pointerleave"] == event_types
-    assert events[3]["type"] == "pointerdown"
-    assert events[3]["pageX"] == pytest.approx(center["x"], abs=1.0)
-    assert events[3]["pageY"] == pytest.approx(center["y"], abs=1.0)
-    assert events[3]["target"] == "pointerArea"
-    assert events[3]["pointerType"] == "pen"
-    # The default value of width and height for mouse and pen inputs is 1
-    assert round(events[3]["width"], 2) == 1
-    assert round(events[3]["height"], 2) == 1
-    assert round(events[3]["pressure"], 2) == 0.36
-    assert events[3]["tiltX"] == -72
-    assert events[3]["tiltY"] == 9
-    assert events[3]["twist"] == 86
-    assert events[6]["type"] == "pointermove"
-    assert events[6]["pageX"] == pytest.approx(center["x"]+10, abs=1.0)
-    assert events[6]["pageY"] == pytest.approx(center["y"]+10, abs=1.0)
-    assert events[6]["target"] == "pointerArea"
-    assert events[6]["pointerType"] == "pen"
-    assert round(events[6]["width"], 2) == 1
-    assert round(events[6]["height"], 2) == 1
-    # The default value of pressure for all inputs is 0.5, other properties are 0
-    assert round(events[6]["pressure"], 2) == 0.5
-    assert events[6]["tiltX"] == 0
-    assert events[6]["tiltY"] == 0
-    assert events[6]["twist"] == 0
-
-
-def test_touch_pointer_properties(session, test_actions_pointer_page, touch_chain):
-    pointerArea = session.find.css("#pointerArea", all=False)
-    center = get_inview_center(pointerArea.rect, get_viewport_rect(session))
-    touch_chain.pointer_move(0, 0, origin=pointerArea) \
-        .pointer_down(width=23, height=31, pressure=0.78, tilt_x=21, tilt_y=-8, twist=355) \
-        .pointer_move(10, 10, origin=pointerArea, width=39, height=35, pressure=0.91, tilt_x=-19, tilt_y=62, twist=345) \
-        .pointer_up() \
-        .pointer_move(80, 50, origin=pointerArea) \
-        .perform()
-    events = get_events(session)
-    assert len(events) == 7
-    event_types = [e["type"] for e in events]
-    assert ["pointerover", "pointerenter", "pointerdown", "pointermove",
-            "pointerup", "pointerout", "pointerleave"] == event_types
-    assert events[2]["type"] == "pointerdown"
-    assert events[2]["pageX"] == pytest.approx(center["x"], abs=1.0)
-    assert events[2]["pageY"] == pytest.approx(center["y"], abs=1.0)
-    assert events[2]["target"] == "pointerArea"
-    assert events[2]["pointerType"] == "touch"
-    assert round(events[2]["width"], 2) == 23
-    assert round(events[2]["height"], 2) == 31
-    assert round(events[2]["pressure"], 2) == 0.78
-    assert events[2]["tiltX"] == 21
-    assert events[2]["tiltY"] == -8
-    assert events[2]["twist"] == 355
-    assert events[3]["type"] == "pointermove"
-    assert events[3]["pageX"] == pytest.approx(center["x"]+10, abs=1.0)
-    assert events[3]["pageY"] == pytest.approx(center["y"]+10, abs=1.0)
-    assert events[3]["target"] == "pointerArea"
-    assert events[3]["pointerType"] == "touch"
-    assert round(events[3]["width"], 2) == 39
-    assert round(events[3]["height"], 2) == 35
-    assert round(events[3]["pressure"], 2) == 0.91
-    assert events[3]["tiltX"] == -19
-    assert events[3]["tiltY"] == 62
-    assert events[3]["twist"] == 345
-
-
 @pytest.mark.parametrize("drag_duration", [0, 300, 800])
 @pytest.mark.parametrize("dx, dy", [
     (20, 0), (0, 15), (10, 15), (-20, 0), (10, -15), (-10, -15)
